+++ packages/protocol/test/layer1/shasta/inbox/InboxChainAdvancement.t.sol
@@ -421,7 +421,7 @@ contract InboxChainAdvancement is InboxTest {
             LibBlobs.BlobReference memory proposalBlobRef = createValidBlobReference(i);
             IInbox.ClaimRecord[] memory emptyClaimRecords = new IInbox.ClaimRecord[](0);
             bytes memory proposalData =
-                abi.encode(uint64(0), proposalCoreState, proposalBlobRef, emptyClaimRecords);
+                abi.encode(uint48(0), proposalCoreState, proposalBlobRef, emptyClaimRecords);
 
             vm.startPrank(Alice);
             setupBlobHashes();
@@ -489,7 +489,7 @@ contract InboxChainAdvancement is InboxTest {
 
         LibBlobs.BlobReference memory blobRef = createValidBlobReference(numProposals + 1);
 
-        bytes memory proposeData = abi.encode(uint64(0), coreState, blobRef, claimRecords);
+        bytes memory proposeData = abi.encode(uint48(0), coreState, blobRef, claimRecords);
 
         mockProposerAllowed(Carol);
         mockForcedInclusionDue(false);
