+++ packages/protocol/test/layer1/shasta/inbox/InboxFinalization.t.sol
@@ -237,7 +237,7 @@ contract InboxFinalization is InboxTest {
         LibBlobs.BlobReference memory blobRef =
             LibBlobs.BlobReference({ blobStartIndex: 1, numBlobs: 1, offset: 0 });
         bytes memory data =
-            encodeProposalDataWithProposals(uint64(0), coreState, proposals, blobRef, claimRecords);
+            encodeProposalDataWithProposals(uint48(0), coreState, proposals, blobRef, claimRecords);
 
         // Submit proposal
         vm.prank(Alice);
